Zindex works

+++ src/components/transform.hpp
@@ -4,12 +4,12 @@
 
 namespace cmpt {
 	struct Transform {
-		Transform(glm::vec2 position = glm::vec2(0.0f), glm::vec2 scale = glm::vec2(1.0f), float rotation = 0.0f, int zIndex = 0)
+		Transform(glm::vec2 position = glm::vec2(0.0f), glm::vec2 scale = glm::vec2(1.0f), float rotation = 0.0f, float zIndex = 0.0f)
 		: position(position), scale(scale), rotation(rotation), zIndex(zIndex) {}
 
 		glm::vec2 position;
 		glm::vec2 scale;
 		float rotation;
-		int zIndex;
+		float zIndex;
 	};
 }
